fix: disable async background flush of traces
When running in complex test suite that mocks many things, trying to
flush in the background traces does not work. For example network
requests might be mocked, or time might be mocked, making SSL connection
impossible.

This stores the span in memory until a flush is requested at the end of
pytest.

This also means there is no way to capture the opentelemetry log in the
background are we can now intercept any error raised by the code while
flushing the traces.

Change-Id: I18bbea366996f8b7f49a8cca409a91b0b1481aff