fix: store only test names retrieved from quarantined endpoint (#182)
The quarantine endpoint returns a list of dict, with various data about
each quarantined tests. The quarantine stored those dict and tried to
compare each test name to the dict retrieved, which caused the
quarantine to not detect any test as quarantined.

If needed in the future, we could still use dict data from each test retrieve from the endpoint, but for now we just need the test names.

Fixes MRGFY-5846