chore: Handle `pytest-timeout`
The goal of this change is to avoid reaching a timeout configured with
`pytest-timeout`. Indeed, the flaky detection mechanism can allocate a
large amount of time for a single test and the timeout will be
triggered. To avoid this, the timeout is now taken into account. The
next change will add this information in the report generated by the
flaky detection.

References: MRGFY-6110
Change-Id: I6bebcffcee617f2bb0c77ec54b6e6204a2eb1daf