fix(ci-insights): Compute a proper per-test deadline
The goal of this change is to compute a deadline for the flaky detection
for each test. Instead of having only a global deadline used to stop if
we're exceeding the overall budget, we'll now choose between 3
parameters:
1. The global deadline (as before),
2. A deadline computed based on the initial run (with some variability),
3. A deadline determined from the pytest-timeout plugin.
The goal is to really stop using too much CI time.

References: MRGFY-6172
Change-Id: I461d11856688b18cad3d999ca401f87feee579db