fix(ci-insights): Use pytest-timeout as a deadline
The goal of this change is to use the timeout as a deadline when
rerunning tests instead of using it to reduce the number of reruns.
Thus, if a rerun take longer to execute, the deadline will work as
expected. We now have a unified behavior for the global and the
pytest-timeout deadlines.

References: MRGFY-6172
Change-Id: Ifa12e63dce016a864463d1696019b78824bec4c8