fix(ci-insights): Increase timeout margin when rerunning tests
The goal of this change is to use a margin of 25% instead of 10% to
really make sure that we are not going to crash the CI with our flaky
detection mechanism. The goal is to reduce the capabilities of the flaky
detection, but inform the user how to optimize it in the report instead
of breaking the CI. The real fix for the client is to use the `func_only`
option of pytest-timeout.

Fixes: MRGFY-6282
Change-Id: If8cd29f35f9ded3bd3cad68af869fe326e244fc9