fix(ci-insights): Properly detect if we can rerun a test
The goal of this change is to use the initial duration to detect if we
can really rerun a test entirely. It's to mitigate the issues when a
timeout occurs during a run. Thus, we should not start any new execution
if it's not feasible based on the initial duration of the test.

Fixes: MRGFY-6282
Change-Id: I1652e7ffd156cbc4a4ec080bb7aa30094cc27a4e